requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atlassian-python-api==4.0.4
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
Deprecated==1.2.18
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jmespath==1.0.1
markdown-it-py==3.0.0
mcp==1.9.1
mdurl==0.1.2
oauthlib==3.2.2
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.5
starlette==0.46.2
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wrapt==1.17.2
